requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pymysql==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
gunicorn==21.2.0
